deletes middleware package entirely
